requirements.txt
azure-functions==1.3.0
Werkzeug==0.16.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
flask-restplus==0.12.1
azf-wsgi==0.3.1